[refactor] user view refoctoring
